+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,75 @@
 # @unipasswallet/wallet
 
+## 0.0.1-alpha.32
+
+### Patch Changes
+
+- @unipasswallet/wallet: fix base64url for browser
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.32
+  - @unipasswallet/dkim-base@0.0.1-alpha.32
+  - @unipasswallet/keys@0.0.1-alpha.32
+  - @unipasswallet/network@0.0.1-alpha.32
+  - @unipasswallet/relayer@0.0.1-alpha.27
+  - @unipasswallet/transactions@0.0.1-alpha.31
+  - @unipasswallet/utils@0.0.1-alpha.32
+
+## 0.0.1-alpha.31
+
+### Patch Changes
+
+- unipasswallet/keys: fix typo
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.31
+  - @unipasswallet/dkim-base@0.0.1-alpha.31
+  - @unipasswallet/keys@0.0.1-alpha.31
+  - @unipasswallet/network@0.0.1-alpha.31
+  - @unipasswallet/relayer@0.0.1-alpha.26
+  - @unipasswallet/transactions@0.0.1-alpha.30
+  - @unipasswallet/utils@0.0.1-alpha.31
+
+## 0.0.1-alpha.30
+
+### Patch Changes
+
+- @unipasswallet/keys: fix signType for Keyset create
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.30
+  - @unipasswallet/dkim-base@0.0.1-alpha.30
+  - @unipasswallet/keys@0.0.1-alpha.30
+  - @unipasswallet/network@0.0.1-alpha.30
+  - @unipasswallet/relayer@0.0.1-alpha.25
+  - @unipasswallet/transactions@0.0.1-alpha.29
+  - @unipasswallet/utils@0.0.1-alpha.30
+
+## 0.0.1-alpha.29
+
+### Patch Changes
+
+- @unipasswallet/keys: add updateSignType for KeyOpenIDWithEmail
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.29
+  - @unipasswallet/dkim-base@0.0.1-alpha.29
+  - @unipasswallet/keys@0.0.1-alpha.29
+  - @unipasswallet/network@0.0.1-alpha.29
+  - @unipasswallet/relayer@0.0.1-alpha.24
+  - @unipasswallet/transactions@0.0.1-alpha.28
+  - @unipasswallet/utils@0.0.1-alpha.29
+
+## 0.0.1-alpha.28
+
+### Patch Changes
+
+- #58: add openID validate
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.28
+  - @unipasswallet/dkim-base@0.0.1-alpha.28
+  - @unipasswallet/keys@0.0.1-alpha.28
+  - @unipasswallet/network@0.0.1-alpha.28
+  - @unipasswallet/relayer@0.0.1-alpha.23
+  - @unipasswallet/transactions@0.0.1-alpha.27
+  - @unipasswallet/utils@0.0.1-alpha.28
+
 ## 0.0.1-alpha.27
 
 ### Patch Changes